Rust debugging scene overview
I went quite a bit deeper with research in order to give the definitive
answer to age-old question: What Debugger extension to use for Rust on
VSCode.

The whole debugging story is fascinating and I've included some
background information for people to have a better understanding of:
  - why some Rust values are not shown properly in debuggers
  - why you can't put a breakpoint on some lines of code
  - why some debuggers are better or worse than others

I want to come up with an example program that shows interesting aspects
of Rust debugging and showcases some limitations, too. But for now it's
just an info dump of theoretical information.