The emoji example didn't compile because of a hidden character.
The editor used to create the file inserted an unprintable extra Unicode Scalar Value - U+FEOF. This is a 'Variation Selector' and it tells the UI to render the previous Code Point as an 'emoji' as opposed to rendering it as 'text'. Unfortunately, this mean the single-quote marks contained two Unicode Scalar Values instead of one, and we consequently got an error from rustc.

This change removes the U+FEOF. The emoji looks rubbish in comparison, but the code does not compile. Or at least, it fails where the comments say it should fail, as opposed to failing earlier on.

See https://unicode-table.com/en/FE0F/