extended async await intro slides
I gave a presentation for a training yesterday based on an earlier iteration
of these slides, and it went pretty well. Mind you, we had about an hour to go though it.

The material is very condensed, but I had a goal of explaining why things are
the way they are.

Highlights:
  * lazy `Future`s
  * `async fn main()`
  * single-threaded vs multithreaded executors
  * why are there different runtimes for async
  * common gotcha around `Mutex`
  * async iteration with `while let await`
  * async traits and why we want them so much
  * why do libraries have to hard code runtime support