Lifetimes with Serde assignment
This assignment starts with JSON deserializing into a struct with own
`String` and gradually move towards use of `Cow`.

Lifetimes show up once, but the initial code is crafted in such a way
that you have to act on your reference potentially outliving
the underlying owned `String`.

The assignment is very quick to complete with most lines of code either
put directly into the text itself or linked from the description.
However, there's a potential of a good discussion, and to facilitate it
I've put a lift of questions at the end.